% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressur-package.R
\docType{package}
\name{pressur}
\alias{pressur}
\alias{pressur-package}
\title{Query and Orchestrate the 'WordPress' 'API'}
\description{
'WordPress' has a fairly comprehensive 'API' <https://developer.wordpress.com/>
that makes it possible to perform blog orchestration ('CRUD' operations on posts, users,
sites, etc.) as well as retrieve and process blog statistics. Tools are provided to work
with the 'WordPress' 'API' functions.
}
\author{
Bob Rudis (bob@rud.is)
}
