% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about-me.R
\name{wp_about_me}
\alias{wp_about_me}
\title{Get your user information}
\usage{
wp_about_me()
}
\description{
Get your user information
}
\examples{
\dontrun{
wp_auth()
wp_about_me()
}
}
\references{
\url{https://developer.wordpress.com/docs/api/1.1/get/me/}
}
