% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-posts.R
\name{wp_get_my_posts}
\alias{wp_get_my_posts}
\title{Get all 'my' posts across all sites}
\usage{
wp_get_my_posts(.quiet = FALSE)
}
\arguments{
\item{.quiet}{if \code{TRUE} then no progress information will be displayed}
}
\description{
Get all 'my' posts across all sites
}
\examples{
\dontrun{
wp_auth()
my_posts <- wp_get_my_posts()
}
}
\references{
\url{https://developer.wordpress.com/docs/api/1.1/get/me/posts/}
}
