% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-stats.R
\name{wp_post_stats}
\alias{wp_post_stats}
\title{Retrieve statistics for a WordPress post}
\usage{
wp_post_stats(site, post_id)
}
\arguments{
\item{site}{site id or domain}

\item{post_id}{a valid post id}
}
\description{
TODO: make most of the list columns more immediately usable
}
\note{
I've only had this work successfully with my blog by using the site id.
}
\examples{
\dontrun{
wp_auth()
me <- wp_about_me()
wp_post_stats(me$primary_blog, "7713")
}
}
\references{
\url{https://developer.wordpress.com/docs/api/1.1/get/sites/$site/stats/post/$post_id/}
}
