% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-info.R
\name{wp_site_info}
\alias{wp_site_info}
\title{Get information about a site}
\usage{
wp_site_info(site)
}
\arguments{
\item{site}{site id or domain}
}
\description{
Get information about a site
}
\note{
I've only had this work successfully with my blog by using the site id.
}
\examples{
\dontrun{
wp_auth()
me <- wp_about_me()
wp_site_info(me$primary_blog)
}
}
\references{
\url{https://developer.wordpress.com/docs/api/1.2/get/sites/$site/}
}
