% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-stats.R
\name{wp_site_stats}
\alias{wp_site_stats}
\title{Get a site's stats}
\usage{
wp_site_stats(site)
}
\arguments{
\item{site}{site id or domain}
}
\description{
Get a site's stats
}
\note{
I've only had this work successfully with my blog by using the site id.
}
\examples{
\dontrun{
wp_auth()
me <- wp_about_me()
wp_site_stats(me$primary_blog)
}
}
\references{
\url{https://developer.wordpress.com/docs/api/1.1/get/sites/$site/stats/}
}
