% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about-me.R
\name{wp_about_me}
\alias{wp_about_me}
\title{Get metadata about the current user.}
\usage{
wp_about_me()
}
\value{
list with user metadata
}
\description{
Get metadata about the current user.
}
\examples{
if (interactive()) {
  wp_auth()
  wp_about_me()
}
}
\references{
<https://developer.wordpress.com/docs/api/1.1/get/me/>
}
