% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_auth.R
\name{wp_auth}
\alias{wp_auth}
\title{Authenticate to WordPress}
\usage{
wp_auth()
}
\description{
Call this at the start of any WordPress API interaction
}
\examples{
\dontrun{
wp_auth()
}
}
\references{
\url{https://developer.wordpress.com/docs/oauth2/}
}
