% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-info.R
\name{wp_site_info}
\alias{wp_site_info}
\title{Get information about a site}
\usage{
wp_site_info(site)
}
\arguments{
\item{site}{site id or domain; if not specified, the primary site of the
authenticated user will be used.}
}
\description{
Get information about a site
}
\examples{
if (interactive() {
  wp_auth()
  wp_site_info()
}
}
\references{
<https://developer.wordpress.com/docs/api/1.2/get/sites/$site/>
}
