% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-stats.R
\name{wp_site_stats}
\alias{wp_site_stats}
\title{Get a site's stats}
\usage{
wp_site_stats(site)
}
\arguments{
\item{site}{site id or domain; if not specified, the primary site of the
authenticated user will be used.}
}
\value{
list with a great deal of stats metadata. You are probably most
        interested in the `visits` element.
}
\description{
Get a site's stats
}
\examples{
if (interactive()) {
  wp_auth()
  wp_site_stats()
}
}
\references{
<https://developer.wordpress.com/docs/api/1.1/get/sites/$site/stats/>
}
